chore(nlu-engine): refactor mltoolkit so no unused references are kept on model
