+++ packages/nlu-engine/src/engine/predict-pipeline.ts
@@ -7,7 +7,7 @@ import {
   IntentPrediction as StanIntentPrediction,
   PredictOutput
 } from 'src/typings'
-import { MLToolkit } from '../ml/typings'
+import * as MLToolkit from '../ml/toolkit'
 
 import { CustomEntityExtractor } from './entities/custom-extractor'
 import { IntentPrediction, IntentPredictions, NoneableIntentPredictions } from './intents/intent-classifier'
