+++ packages/nlu-engine/src/engine/test-utils/fake-kmeans.ts
@@ -1,4 +1,4 @@
-import { MLToolkit } from '../../ml/typings'
+import * as MLToolkit from '../../ml/toolkit'
 
 export const fakeKmeans: typeof MLToolkit.KMeans = {
   kmeans: (data: MLToolkit.KMeans.DataPoint[], k: number, options: MLToolkit.KMeans.KMeansOptions) => {
