+++ packages/nlu-engine/src/engine/test-utils/fake-svm.ts
@@ -1,18 +1,17 @@
 import _ from 'lodash'
-import { MLToolkit } from '../../ml/typings'
+import * as MLToolkit from '../../ml/toolkit'
+import { Logger } from '../../typings'
 
-export class FakeSvmTrainer implements MLToolkit.SVM.Trainer {
-  private _isTrained = false
-  constructor() {}
-  public async train(
+export class FakeSvmTrainer {
+  public static async train(
     points: MLToolkit.SVM.DataPoint[],
-    options?: MLToolkit.SVM.SVMOptions | undefined,
-    callback?: MLToolkit.SVM.TrainProgressCallback | undefined
+    options: MLToolkit.SVM.SVMOptions | undefined,
+    logger: Logger,
+    callback: MLToolkit.SVM.TrainProgressCallback | undefined
   ): Promise<Buffer> {
     if (!points.length) {
       throw new Error('fake SVM needs datapoints')
     }
-    this._isTrained = true
     callback?.(1)
     return Buffer.from(
       _(points)
@@ -22,20 +21,19 @@ export class FakeSvmTrainer implements MLToolkit.SVM.Trainer {
         .join(',')
     )
   }
-  public isTrained(): boolean {
-    return this._isTrained
-  }
 }
 
-export class FakeSvmPredictor implements MLToolkit.SVM.Predictor {
+export class FakeSvmPredictor implements MLToolkit.SVM.IPredictor {
   constructor(private model: Buffer) {}
+
+  public static async create(model: Buffer) {
+    return new FakeSvmPredictor(model)
+  }
 
   public async predict(coordinates: number[]): Promise<MLToolkit.SVM.Prediction[]> {
     const labels = this.model.toString().split(',')
     return labels.map((label) => ({ label, confidence: 1 / labels.length }))
   }
-
-  public async initialize() {}
 
   public isLoaded(): boolean {
     return true
