+++ packages/nlu-engine/src/engine/training-pipeline.ts
@@ -1,6 +1,6 @@
 import Bluebird from 'bluebird'
 import _ from 'lodash'
-import { MLToolkit } from '../ml/typings'
+import * as MLToolkit from '../ml/toolkit'
 import { Logger } from '../typings'
 
 import { watchDog } from '../utils/watch-dog'
