+++ packages/nlu-engine/src/ml/sentencepiece/index.ts
@@ -1,6 +1,1 @@
-import { makeProcessor } from '@botpress/node-sentencepiece'
-import { MLToolkit } from '../typings'
-
-export const processor: () => Promise<MLToolkit.SentencePiece.Processor> = () => {
-  return makeProcessor()
-}
+export { makeProcessor as createProcessor, Processor } from '@botpress/node-sentencepiece'
