+++ packages/nlu-engine/src/ml/svm/svm.test.ts
@@ -1,30 +1,28 @@
 import { Logger } from 'src/typings'
-import { MLToolkit } from '../typings'
-import { Predictor, Trainer } from '.'
+import { DataPoint, Predictor, Trainer } from '.'
 
 const SEED = 42
 
-/**
- * WARNING:
- *  If test fails it may be because of your Linux distribution.
- *  Try editing 'jest-before.ts' file your distribution.
- */
+const dummyLogger: Partial<Logger> = { debug: () => {} }
+const dummyCallback = () => {}
+
 describe('SVM', () => {
   test('Trainer should work with basic problems', async () => {
-    // prettier-ignore
-    const line: MLToolkit.SVM.DataPoint[] = [
+    const line: DataPoint[] = [
       { coordinates: [0, 0], label: 'A' },
       { coordinates: [0, 1], label: 'A' },
       { coordinates: [1, 0], label: 'B' },
       { coordinates: [1, 1], label: 'B' }
     ]
 
-    const dummyLogger: Partial<Logger> = { debug: () => {} }
-    const trainer = new Trainer(dummyLogger as Logger)
-    const mod = await trainer.train(line, { classifier: 'C_SVC', kernel: 'LINEAR', c: 1, seed: SEED })
+    const mod = await Trainer.train(
+      line,
+      { classifier: 'C_SVC', kernel: 'LINEAR', c: 1, seed: SEED },
+      dummyLogger as Logger,
+      dummyCallback
+    )
 
-    const predictor = new Predictor(mod)
-    await predictor.initialize()
+    const predictor = await Predictor.create(mod)
 
     const r1 = await predictor.predict([0, 0])
     const r2 = await predictor.predict([1, 1])
@@ -38,20 +36,21 @@ describe('SVM', () => {
   })
 
   test('Trainer should throw when vectors have different lengths', async () => {
-    // prettier-ignore
-    const line: MLToolkit.SVM.DataPoint[] = [
+    const line: DataPoint[] = [
       { coordinates: [0, 0, 0], label: 'A' },
       { coordinates: [0, 1], label: 'A' },
       { coordinates: [1, 0], label: 'B' },
       { coordinates: [1, 1], label: 'B' }
     ]
 
-    const dummyLogger: Partial<Logger> = { debug: () => {} }
-    const trainer = new Trainer(dummyLogger as Logger)
-
     let errorThrown = false
     try {
-      await trainer.train(line, { classifier: 'C_SVC', kernel: 'LINEAR', c: [1], seed: SEED })
+      await Trainer.train(
+        line,
+        { classifier: 'C_SVC', kernel: 'LINEAR', c: [1], seed: SEED },
+        dummyLogger as Logger,
+        dummyCallback
+      )
     } catch (err) {
       errorThrown = true
     }
